rename for consistancy, more documentation, update Changelog
