+++ README.md
@@ -682,7 +682,7 @@ corporation.corporation_url # => "http://"
 ```ruby
 options = { token: 'token123', character_id: 90729314 }
 
-character_jobs = EveOnline::ESI::CharacterIndustryJob.new(options)
+character_jobs = EveOnline::ESI::CharacterIndustryJobs.new(options)
 
 character_jobs.scope => # => "esi-industry.read_character_jobs.v1"
 
@@ -690,28 +690,72 @@ character_jobs.jobs.size # => 3
 
 job = character_jobs.jobs.first
 
-job.as_json = # => {:activity_id => 5,
-              #     :blueprint_id => 1024839597103,
-              #     :blueprint_location_id => 1023579231924,
-              #     :blueprint_type_id => 28607,
-              #     :completed_character_id => nil,
-              #     :completed_date => nil,
-              #     :cost => 901631.0,
-              #     :duration => 625697,
-              #     :end_date => Sat, 18 Nov 2017 02:20:24 UTC +00:00,
-              #     :facility_id => 1023579231924,
-              #     :installer_id => 93997721,
-              #     :job_id => 344003357,
-              #     :licensed_runs => 1,
-              #     :output_location_id => 1023579231924,
-              #     :pause_date => nil,
-              #     :probability => 1.0,
-              #     :product_type_id => 28607,
-              #     :runs => 2,
-              #     :start_date => Fri, 10 Nov 2017 20:32:07 UTC +00:00,
-              #     :station_id => 1023579231924,
-              #     :status => "active",
-              #     :successful_runs => nil}
+job.as_json = # => {:activity_id=>5,
+              #     :blueprint_id=>1024839597103,
+              #     :blueprint_location_id=>1023579231924,
+              #     :blueprint_type_id=>28607,
+              #     :completed_character_id=>nil,
+              #     :completed_date=>nil,
+              #     :cost=>902034.0,
+              #     :duration=>625697,
+              #     :end_date=>Sat, 25 Nov 2017 16:04:31 UTC +00:00,
+              #     :facility_id=>1023579231924,
+              #     :installer_id=>93997721,
+              #     :job_id=>344732396,
+              #     :licensed_runs=>1,
+              #     :output_location_id=>1023579231924,
+              #     :pause_date=>nil,
+              #     :probability=>1.0,
+              #     :product_type_id=>28607,
+              #     :runs=>2,
+              #     :start_date=>Sat, 18 Nov 2017 10:16:14 UTC +00:00,
+              #     :station_id=>1023579231924,
+              #     :status=>"active",
+              #     :successful_runs=>nil}
+
+job.activity_id # => 5
+
+job.blueprint_id # => 1024839597103
+
+job.blueprint_location_id # => 1023579231924
+
+job.blueprint_type_id # => 28607
+
+job.completed_character_id # => nil
+
+job.completed_date # => nil
+
+job.cost # => 902034.0
+
+job.duration # => 625697
+
+job.end_date # => Sat, 25 Nov 2017 16:04:31 UTC +00:00
+
+job.facility_id # => 1023579231924
+
+job.installer_id # => 93997721
+
+job.job_id # => 344732396
+
+job.licensed_runs # => 1
+
+job.output_location_id # => 1023579231924
+
+job.pause_date # => nil
+
+job.probability # => 1.0
+
+job.product_type_id # => 28607
+
+job.runs # => 2
+
+job.start_date # => Sat, 18 Nov 2017 10:16:14 UTC +00:00
+
+job.station_id # => 1023579231924
+
+job.status # => "active"
+
+job.successful_runs # => nil
 
 # TODO: add pagination support
 
@@ -722,7 +766,7 @@ job.as_json = # => {:activity_id => 5,
 ```ruby
 options = { token: 'token123', corporation_id: 98146630 }
 
-corporation_jobs = EveOnline::ESI::CorporationIndustryJob.new(options)
+corporation_jobs = EveOnline::ESI::CorporationIndustryJobs.new(options)
 
 corporation_jobs.scope # => "esi-industry.read_corporation_jobs.v1"
 
@@ -730,28 +774,72 @@ corporation_jobs.jobs.size # => 23
 
 job = corporation_jobs.jobs.first
 
-job.as_json # => {:activity_id => 1,
-            #     :blueprint_id => 1025589624582,
-            #     :blueprint_location_id => 1024956764558,
-            #     :blueprint_type_id => 28607,
-            #     :completed_character_id => nil,
-            #     :completed_date => nil,
-            #     :cost => 18131639.0,
-            #     :duration => 424816,
-            #     :end_date => Wed, 15 Nov 2017 18:14:02 UTC +00:00,
-            #     :facility_id => 1023579231924,
-            #     :installer_id => 93174304,
-            #     :job_id => 344001862,
-            #     :licensed_runs => 1,
-            #     :output_location_id => 1024956764558,
-            #     :pause_date => nil,
-            #     :probability => 1.0,
-            #     :product_type_id => 28606,
-            #     :runs => 1,
-            #     :start_date => Fri, 10 Nov 2017 20:13:46 UTC +00:00,
-            #     :station_id => nil,
-            #     :status => "active",
-            #     :successful_runs => nil}
+job.as_json # => {:activity_id=>1,
+            #     :blueprint_id=>1026042055832,
+            #     :blueprint_location_id=>1024956764558,
+            #     :blueprint_type_id=>28607,
+            #     :completed_character_id=>nil,
+            #     :completed_date=>nil,
+            #     :cost=>19505804.0,
+            #     :duration=>424816,
+            #     :end_date=>Thu, 23 Nov 2017 09:20:30 UTC +00:00,
+            #     :facility_id=>1023579231924,
+            #     :installer_id=>93174304,
+            #     :job_id=>344736432,
+            #     :licensed_runs=>1,
+            #     :output_location_id=>1024956764558,
+            #     :pause_date=>nil,
+            #     :probability=>1.0,
+            #     :product_type_id=>28606,
+            #     :runs=>1,
+            #     :start_date=>Sat, 18 Nov 2017 11:20:14 UTC +00:00,
+            #     :station_id=>nil,
+            #     :status=>"active",
+            #     :successful_runs=>nil}
+
+job.activity_id # => 1
+
+job.blueprint_id # => 1026042055832
+
+job.blueprint_location_id # => 1024956764558
+
+job.job.blueprint_type_id # => 28607
+
+job.completed_character_id # => nil
+
+job.completed_date # => nil
+
+job.cost # => 19505804.0
+
+job.duration # => 424816
+
+job.end_date # => Thu, 23 Nov 2017 09:20:30 UTC +00:00
+
+job.facility_id # => 1023579231924
+
+job.installer_id # => 93174304
+
+job.job_id # => 344736432
+
+job.licensed_runs # => 1
+
+job.output_location_id # => 1024956764558
+
+job.pause_date # => nil
+
+job.probability # => 1.0
+
+job.product_type_id # => 28606
+
+job.runs # => 1
+
+job.start_date # => Sat, 18 Nov 2017 11:20:14 UTC +00:00
+
+job.station_id # => nil
+
+job.status # => "active"
+
+job.successful_runs # => nil
 
 # TODO: add pagination support
 
@@ -1353,6 +1441,7 @@ Issue reports and pull requests are welcome on GitHub at https://github.com/biow
 ## Contributors. Thank you everyone!
 
 * Ian Flynn (@monban)
+* Mekaret Eriker (@Mekaret)
 
 ## License
