+++ lib/eve_online.rb
@@ -39,7 +39,7 @@ require 'eve_online/esi/character_loyalty_points'
 require 'eve_online/esi/character_wallet'
 require 'eve_online/esi/character_assets'
 require 'eve_online/esi/character_attributes'
-require 'eve_online/esi/character_industry_job'
+require 'eve_online/esi/character_industry_jobs'
 require 'eve_online/esi/character_clones'
 require 'eve_online/esi/character_implants'
 require 'eve_online/esi/character_blueprints'
@@ -47,7 +47,7 @@ require 'eve_online/esi/character_standings'
 require 'eve_online/esi/character_fatigue'
 
 require 'eve_online/esi/corporation'
-require 'eve_online/esi/corporation_industry_job'
+require 'eve_online/esi/corporation_industry_jobs'
 
 require 'eve_online/esi/alliance'
 require 'eve_online/esi/alliance_icon'
@@ -62,7 +62,7 @@ require 'eve_online/esi/models/skill'
 require 'eve_online/esi/models/asset'
 require 'eve_online/esi/models/attributes'
 require 'eve_online/esi/models/character'
-require 'eve_online/esi/models/industry_jobs'
+require 'eve_online/esi/models/industry_job'
 require 'eve_online/esi/models/character_portrait'
 require 'eve_online/esi/models/blueprint'
 require 'eve_online/esi/models/standing'
