+++ lib/eve_online/esi/models/industry_job.rb
